+++ external/sparse_linear_solver/linear_solver.h
@@ -35,28 +35,28 @@ extern "C" {
 typedef struct LinearSolver LinearSolver;
 
 LinearSolver *EIG_linear_solver_new(
-  int num_rows,
-  int num_columns,
-  int num_right_hand_sides);
+  unsigned long num_rows,
+  unsigned long num_columns,
+  unsigned long num_right_hand_sides);
 
 LinearSolver *EIG_linear_least_squares_solver_new(
-  int num_rows,
-  int num_columns,
-  int num_right_hand_sides);
+  unsigned long num_rows,
+  unsigned long num_columns,
+  unsigned long num_right_hand_sides);
 
 void EIG_linear_solver_delete(LinearSolver *solver);
 
 /* Variables (x). Any locking must be done before matrix construction. */
 
-void EIG_linear_solver_variable_set(LinearSolver *solver, int rhs, int index, double value);
-double EIG_linear_solver_variable_get(LinearSolver *solver, int rhs, int index);
-void EIG_linear_solver_variable_lock(LinearSolver *solver, int index);
-void EIG_linear_solver_variable_unlock(LinearSolver *solver, int index);
+void EIG_linear_solver_variable_set(LinearSolver *solver, unsigned long rhs, unsigned long index, double value);
+double EIG_linear_solver_variable_get(LinearSolver *solver, unsigned long rhs, unsigned long index);
+void EIG_linear_solver_variable_lock(LinearSolver *solver, unsigned long index);
+void EIG_linear_solver_variable_unlock(LinearSolver *solver, unsigned long index);
 
 /* Matrix (A) and right hand side (b) */
 
-void EIG_linear_solver_matrix_add(LinearSolver *solver, int row, int col, double value);
-void EIG_linear_solver_right_hand_side_add(LinearSolver *solver, int rhs, int index, double value);
+void EIG_linear_solver_matrix_add(LinearSolver *solver, unsigned long row, unsigned long col, double value);
+void EIG_linear_solver_right_hand_side_add(LinearSolver *solver, unsigned long rhs, unsigned long index, double value);
 
 /* Solve. Repeated solves are supported, by changing b between solves. */
